Add isolation flag (aka option) --isolate(-i).
When --isolate(-i) option is used test matched to regular expression in
--single(-s) option will be launched in dedicated process.

For isolated tests additional process will be created. For example, if
you have 5 tests and 4 cores, without isolation tests will run in for
processes, but with isolation when one of the test matched to isolation
pattern parallel_test forks 5 processes. This was made intentionally
because of assumption that most of the tests won't matched to
isolation pattern. So, tests in isolation should finished earlier then
rest of tests.
I assume that isolated tests will be I/O intensive or just wait for
some time in timeout. So, additional process shouldn't make a big
overhead.

Changes:
- Make --isolate(-i) option a flag
- Change description of --single(-s) and --isolate(-i) options
- Add condition which prevents to iterate through already grouped tests
- Rename method Grouper.isolate! to Grouper.isolated
- Eliminate side effect of Grouper.isolate! which change incoming list
  of tests. Now Grouper.isolated returned array of grouped tests.
- Replace usage of short options for rake tasks to long ones
- Add condition which stops grouping process when there is not free cores
  left.
- Fix and adopt specs to match new functionality
- Add new specs