Fixes how number of running processes are retrieved
Adds a rspec_in_path? method that checks for handling of 'rspec' command without the need for adding a bin or script path

moves to using Process.spawn to handle problem of subprocesses that are started and detached still block completion of the 'open' command

increasing version to 0.16.11

updated rspec/runner_spec.rb tests to ensure handling of new rspec_in_path? method