Add sleep after stdout flush
Unfortunately, just calling flush is not a guarantee that stdout will be written. 

The proper way to do this is probably to do all writing from a single thread and push strings through a queue to it, but this is a simple solution that should work most of the time.