Fix for "Pass files or folders to run"
When running

```
bundle exec rake "parallel:spec[7, ^spec/features, --rutabaga -- -r rutabaga --]"
```

for some reason the directory to run against wasn't being passed correctly and error

```
Pass files or folders to run
```

was given. This appears to be because the directory was the first argument when
it should be the last argument.