Use ruby binary prefix only for bundler stubs
It turns out that using path to binaries relative to the source
code is not the best practice. On Windows rubygems generates .bat
files for every binary during gem installing. However it doesn't
help if we try to locate and run binary relative to the source
code.

This change removes all such usages, so running parallel_*
binaries will work on Windows and on Linux the same way. The only
place we still need prefixing with ruby.exe is bundler's binstubs.