Use ParallelTests.with_ruby_binary only for bin/parallel_* binaries (#590)
* Use ParallelTests.with_ruby_binary only for bin/parallel_* binaries

This change fixes rake tasks on Windows. The only place where adding
ruby binary is necessary for bin/parallel_* binaries. For all other
cases it may leads to some ambiguous command lines like
`ruby.exe ruby rake` which don't work.

* Use ruby binary prefix only for bundler stubs

It turns out that using path to binaries relative to the source
code is not the best practice. On Windows rubygems generates .bat
files for every binary during gem installing. However it doesn't
help if we try to locate and run binary relative to the source
code.

This change removes all such usages, so running parallel_*
binaries will work on Windows and on Linux the same way. The only
place we still need prefixing with ruby.exe is bundler's binstubs.

* Fix tests

* Don't use Shellwords.escpace where it's not necessary

* Restore using relative path during call parallel_* binaries

* Revert "Fix tests"

This reverts commit a6c3137