Add new command-line options to provide finer control of verbose output.
--verbose currently displays both the command to be executed by each parallel
process before the tests run, and, on failure of a process, the command
needed to rerun the failing process.

We prefer to display only the latter.

This patch adds flags that can be used alone or in combination with
--verbose to fine-tune the output.

--verbose with --no-report-executed-command will show the rerun command
as well as any other output that --verbose controls, but will not
display the per-process commands before the tests start.

--report-executed-command without --verbose will show only the
per-process commands.

--verbose with --no-report-rerun-command will show the per-process
commands and any other output that --verbose controls, but will not show
the rerun commands when tests fail.

--report-rerun-command without --verbose will show only the rerun
commands.