Add new command-line options to provide finer control of verbose output.
--verbose currently displays both the command to be executed by each parallel
process before the tests run, and, on failure of a process, the command
needed to rerun the failing process.

We prefer to display only the latter.

This patch adds flags that can be used instead of --verbose to enable
either the per-process command, or the rerun command, or both.

--verbose-process-command will show the per-process commands, even if
--verbose is not set, or --quiet is set.

--verbose-rerun-command will show the rerun command, even if --verbose
is not set, or --quiet is set.

If --verbose is set, both types of command will be displayed, regardless
of whether the flags for those commands are set or not.