Colorize summarized RSpec results.
RSpec colorizes its results based on if there are any failures (red), pending (yellow) or none of the above (green).

The method used to summarize the results from the various threads, `summarize_results` was not doing this colorization.

This commit applies the same colorization that RSpec does but does not depend on RSpec::Core for colorization. It utilizes red for failure, yellow for pending and green for success.

- Entry in `CHANGELOG.md` added.
- Specs added for all major scenarios:
  - No failures or pending.
  - Just pending.
  - Failures and pending.
  - Just failures.

Co-authored-by: Michael Grosser <michael@grosser.it>