Adding --retain-exit-status which returns the highest exit status from a test run failure
rather than always exiting with 1 when running in parallel.

The use-case for this is test suites that use exit statuses to communicate additional information besides failure such as marking a suite as unstable.

This treats an exit status of 1 as a generic failure which is the default for most suites, and treats any error above 1 as being preferred when present.

This change is backwards compatible for existing test suites that use parallel_tests.