Changed subprocesses to execute without a shell.
Treat subprocess commands as an array of strings. This avoid the need to
include a shell in the subprocess. Instead, the commands execute
directly. As well, there is less need to parse or format shell
syntax (such as string quoting).

Continue to handle IO redirection through IO.popen.

As a caution, this may be considered a backward incompatible change if
users are relying on injecting shell features through the CLI. I don't
expect this to be an issue in practice, but am raising it just in case.
It would be best to apply a major version bump before release to
maintain semantic versioning.