Simplify use of command "test" in subprocesses (#859)
`|| test 1` is equivalent to `true` which always returns a exit status
of 0. Using true is more explicit. See: https://linux.die.net/man/1/true

`&& test 1` is equivalent to using the exit status of the preceding
command, so just use it and avoid the extra syntax.

`&& test` is equivalent to `false` which always returns a exit status of
1. Using false is more explicit. See: https://linux.die.net/man/1/false