Change --exec argument to run subprocesses without a shell (#857)
* Change --exec argument to run subprocesses without a shell

Continues the work from 5ac27e9f551840722c830cf72bec64a3e2ef325d.

Treat subprocess commands originating from `--exec` as an array of
strings. This avoid the need to include a shell in the subprocess.
Instead, the commands execute directly. As well, there is less need to
parse or format shell syntax (such as string quoting).

This change is a backward incompatible change. When the argument parse
sees the `--exec` argument, all remaining arguments are collected as the
arguments to the subprocess to execute. Previously, it only consumed one
argument, allowing for additional arguments after it. This change should
come with a major version bump.

* Change back to making --exec accept a single string argument

Will not execute with a shell, though.