Install spec/fixtures/**/Gemfile in frozen mode
Through the use of `bundle install` these files change on a test run if
the parallel_tests version changes.

For example, right now parallel_tests is version 3.8.1 but is specified
as 3.8.0 in spec/fixtures/rails70/Gemfile.lock. Running rspec results in
a undesired change to this file. When submitting pull requests, it is
best practice to avoid unrelated changes, so this slightly slows down a
contributor's workflow.

Use the frozen configuration option to ensure that these files are
always up to date. If they are not, CI will fail with a useful message.

See https://bundler.io/man/bundle-config.1.html

> frozen (BUNDLE_FROZEN): Disallow changes to the Gemfile. When the
> Gemfile is changed and the lockfile has not been updated, running
> Bundler commands will be blocked. Defaults to true when --deployment
> is used.