Raise custom error when runtime log is too small
Currently, we raise a generic `RuntimeError` when the runtime log is too
small. This can make it hard for calling code to handle the situation.
The best that can be done is attempting to match on the string error
message.

This change introduces a custom `RuntimeLogTooSmallError` that calling
code can more easily rescue. This will make it easier for builds to log
when this event occurs or automatically retry with a different grouping
strategy.

This is a breaking change for anyone who was relying on rescuing
`RuntimeError`.