Document `--only-group` grouping strategy choice (#870)
The documentation says that parallel_test will attempt to use a runtime
grouping strategy if a runtime file is present and otherwise fall back
to the filesize strategy. This behavior can be overridden with an
explicit `--group-by` flag.

The `--only-group` flag changes this by always running with the filesize
strategy when no explicit strategy is given, even if a runtime file is
present. This behavior happens in `lib/parallel_tests/cli.rb:312`.

I was attempting to rely on the default behavior and also use the
`--only-group` flag and couldn't figure out why my runtime log wasn't
getting used. Only by reading the source did I figure out this
undocumented edge-case in the `--only-group` flag.

This commit adds explicit documentation about this. If you want to use
the `--only-group` flag and a runtime log, you _must_ explicitly set the
`--group-by` flag instead of relying on the default behavior.